TECH: Move mock data to connectors and remove useless modules
