+++ src/react/App.spec.js
