+++ src/react/components/utils.js
@@ -1,4 +1,8 @@
-export const getBackgroundImage = (image, text, gradient='255, 255, 255, 0.8') =>
+export const getBackgroundImage = (
+  image,
+  text,
+  gradient = '255, 255, 255, 0.8'
+) =>
   text
     ? `linear-gradient(rgba(${gradient}),rgba(${gradient})), url("${encodeURI(
         image
