+++ src/react/features/connector/mockResponses/GET_HOME.json
