+++ src/react/features/connector/mockResponses/GET_ROOM.json
