+++ src/react/features/connector/mockResponses/home.json
