+++ src/react/utils/cucumber/support/data/expectLinksWithLocations.js
@@ -1,5 +1,5 @@
 export const expectLinksWithLocations = (idPrefix, id, links) => {
-  const { diograph } = require(`../../../../lib/mockResponses/GET_ROOM.json`)
+  const { diograph } = require(`./src/react/features/connector/mockResponses`)
   const diory = diograph[idPrefix + id]
   const locations = Object.values(diory.links).filter(
     ({ id }) => diograph[id].latitude && diograph[id].longitude
