+++ src/react/utils/cucumber/support/data/expectRoomsInHome.js
@@ -1,4 +1,4 @@
 export const expectRoomsInHome = rooms => {
-  const home = require(`../../../../lib/mockResponses/GET_HOME`)
+  const home = require(`./src/react/features/connector/mockResponses`)
   expect(Object.keys(home.rooms).length).toBe(rooms)
 }
